fix(scoring): computed_score perdia las medias fraccionarias en silencio (mig 018)
`computed_score.raw` nacio `integer not null` (002_user_data.sql:91) cuando el
unico tipo de formula seedeado era `sum`. El 2026-06-12 entraron los dos primeros
instrumentos `mean` —TwIVI y PERMA-Profiler— y `meanFormula` no redondea
(formulas/mean.ts:20), asi que toda media fraccionaria revienta el INSERT con
Postgres 22P02 y el error se traga a proposito (score-session.ts:395).

EL ENUNCIADO CON EL QUE LLEGO ERA FALSO EN SUS DOS MITADES. Se recibio como
"computed_score de PERMA nunca se escribe". Se escribe cuando la media cae
entera —en prod hay 4 usuarios con PERMA 9/9 completo— y TwIVI tambien esta
afectado: tiene 10 reglas `mean` y a una cuenta le faltan 5 de 10.

El defecto real es peor que el enunciado: no es ausencia total, que se ve, sino
perdida PARCIAL que depende de lo que respondio el usuario. `sum` (BFI-2-S,
O*NET-IP-SF) siempre da entero y es inmune. Barrido de prod: 15 filas faltantes
en 2 cuentas de prueba (pruebasgvhv1+r6 PERMA 5/9, +r7 PERMA 3/9 y TwIVI 5/10).
Ningun titular tercero afectado, que es lo que permitio decidir sin backfill.

ALCANCE. No observable en el reporte ni en el teaser: los dos leen
report_snapshot. SI observable en el export ARCO (app/api/me/data/route.ts:139,
COMPL-05, art. 15 Ley 1581), que entrega un array computed_scores poblado pero
incompleto sin declarar cuanto falta. El borrado no se afecta.

POR QUE NINGUN SMOKE LO VIO. Los pares de TwIVI estan 10 apart (items.sql:
1,11 CO | 2,12 TR | ..., control de aquiescencia) y el vector conducido fue el
ciclo 2,4,6,3,5, de periodo 5. Como 10 es multiplo de 5, el item k y el k+10
recibieron siempre el mismo valor: toda suma par, toda media entera. Un vector
que parecia variado era, para esta aritmetica, uniforme.

POR QUE numeric Y NO REDONDEAR. PERMA publica medias de dominio 0-10 y el
centering MRAT de TwIVI opera sobre medias. Redondear guardaria un valor que el
motor nunca calculo y se lo entregaria al titular como su puntaje. `normalized`,
en la misma tabla, ya es numeric.

FALSADO ROJO -> VERDE, con la migracion como unica variable. El test fallo con
22P02 —el mismo codigo que prod— antes de la migracion y paso despues. Afirma
igualdad exacta con 4.5, no "no tiro error", para cerrar tambien la puerta de un
"fix" por Math.round.

CONFIRMACION POSITIVA, porque un 0 tambien seria consistente con "el flujo no
corrio": la suite E2E dejo 27 filas fraccionarias de PERMA (9 dims x 3 proyectos
= la aritmetica exacta de los 9 errores por corrida que reportaba #72) contra 0
en BFI-2-S y O*NET-IP-SF. La primera corrida dio 0 errores y 0 filas —un verde
vacio— por faltarle E2E_LOCAL=1, que CI si prende (ci.yml:163) y que se saltea 9
tests del tramo autenticado.

LO QUE NO CIERRA: el E2E sigue sin cubrir la pata TwIVI, porque su fixture
tambien resulta par-uniforme. Lo cubre el test de regresion, que afirma sobre la
columna y sobre cualquier regla `mean` del seed, no sobre un instrumento.

Docs: ADR-042, [GAP-COMPUTED-SCORE-MEDIAS-DESCARTADAS] cerrado, y las 3 filas
que PM-20 dejo nombradas (BACKLOG :40 y :87 decian 9/9+2/3 con ADR-039 ya en
3/3; :147 estaba mal en el dato). Al corregir la tercera aparecio una correccion
de fondo: /api/respond devuelve dos cuerpos distintos en 500 que significan
cosas opuestas —insert_failed (:228) es la fila NO escrita, internal (:243) es
alcanzable solo tras un upsert exitoso— y la unica reproduccion documentada en
CI trae insert_failed. Abre ademas [GAP-TEASER-COBERTURA-BANDAS] P2.

Verificado como CI (nvm 24, db reset, workers=2, E2E_LOCAL=1): typecheck limpio,
lint 19, unit 490 passed | 21 todo (eran 488, +2 exactos del archivo nuevo),
E2E 97 passed / 2 skipped sobre los 3 proyectos.

+++ db/schema/computed-score.ts
@@ -1,5 +1,5 @@
 // Computed score — scoring output with version trace (Gate 1 trazabilidad).
-import { integer, numeric, pgTable, text, timestamp, uuid } from "drizzle-orm/pg-core";
+import { numeric, pgTable, text, timestamp, uuid } from "drizzle-orm/pg-core";
 import { baremo } from "./baremo";
 import { scoringRule } from "./scoring-rule";
 import { user } from "./user";
@@ -14,7 +14,9 @@ export const computedScore = pgTable("computed_score", {
     .notNull()
     .references(() => scoringRule.id),
   baremoId: uuid("baremo_id").references(() => baremo.id),
-  raw: integer("raw").notNull(),
+  // numeric desde mig 018 (ADR-042): las formulas `mean` (PERMA, TwIVI) publican
+  // medias, y `integer` las rechazaba con 22P02 en silencio.
+  raw: numeric("raw").notNull(),
   normalized: numeric("normalized"),
   band: text("band"),
   scoringVersion: text("scoring_version").notNull(),
